requirements.txt
# AlTi RAG Service Dependencies
# Python 3.11+ recommended

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# RAG Stack (LlamaIndex - legacy)
llama-index>=0.12.0
llama-index-vector-stores-chroma>=0.4.0
llama-index-embeddings-openai>=0.3.0
llama-index-llms-openai>=0.3.0
chromadb>=0.5.0

# LangGraph + LangChain (new agentic RAG)
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
langchain-cohere>=0.3.0          # For reranking
rank-bm25>=0.2.0                  # BM25 retrieval

# Document Loaders
llama-index-readers-file>=0.4.0
pymupdf>=1.24.0          # PDF parsing
openpyxl>=3.1.0          # Excel files
python-docx>=1.1.0       # Word documents
python-pptx>=1.0.0       # PowerPoint files
pandas>=2.2.0            # CSV/Excel data handling

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0            # Async HTTP client
tenacity>=9.0.0          # Retry logic

# Evaluation Harness
click>=8.1.0             # CLI framework
pyyaml>=6.0.0            # YAML parsing for test queries

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
